requirements.txt
anyio==4.2.0
asgiref==3.8.1
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrdict==2.0.1
Babel==2.14.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.zoneinfo==0.2.1
bce-python-sdk==0.9.4
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.8
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Django==4.1.13
django-widget-tweaks==1.5.0
djongo==1.3.6
dnspython==2.6.1
easyocr==1.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.13.3
fire==0.5.0
Flask==3.0.2
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.3.1
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython @ file:///D:/bld/ipython_1680185618122/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///D:/bld/jupyter_core_1669775263822/work
kiwisolver==1.4.5
lanms_neo==1.0.2
lazy_loader==0.3
lmdb==1.4.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mongoengine==0.28.2
mpmath==1.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
onnxruntime==1.17.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddleocr==2.7.0.3
paddlepaddle==2.6.0
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pdf2docx==0.5.8
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
Polygon3==3.0.9.1
premailer==3.10.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.2
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyclipper==1.3.0.post5
pycryptodome==3.20.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pymongo==3.12.3
PyMuPDF==1.20.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
PyWavelets==1.4.1
pywin32==227
PyYAML==6.0.1
pyzmq @ file:///C:/b/abs_89aq69t0up/croot/pyzmq_1705605705281/work
rapidfuzz==3.6.1
rarfile==4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.2.4
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
termcolor==2.4.0
tifffile==2023.7.10
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado @ file:///D:/bld/tornado_1656937958156/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1710254411456/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata==2024.1
urllib3==2.2.0
visualdl==2.5.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
